Lab and Morbidity report query
