+++ apps/modernization-api/src/main/resources/graphql/patient-search.graphqls
@@ -24,7 +24,15 @@ input PersonFilter {
   mortalityStatus: String
   ethnicity: String
   vaccinationId: String
+  morbidityId: String
+  documentId: String
+  stateCaseId: String
+  abcCaseId: String
+  cityCountyCaseId: String
+  notificationId: String
   treatmentId: String
+  labReportId: String
+  accessionNumberId: String
   disableSoundex: Boolean
   recordStatus: [RecordStatus!]!
 }
